[FIX] 테스트를 위한 타입 변경

+++ src/test/java/com/soptie/server/routine/fixture/DailyRoutineFixture.java
@@ -21,7 +21,7 @@ public class DailyRoutineFixture {
 
 	private static DailyThemesResponse.DailyThemeResponse createDailyThemeResponse(int i) {
 		return DailyThemesResponse.DailyThemeResponse.builder()
-				.themeId(i)
+				.themeId((long)i)
 				.name("theme name" + i)
 				.iconImageUrl("https://...")
 				.backgroundImageUrl("https://...")
